refactor: move ga components to features folder
