+++ src/app/layout.tsx
@@ -5,10 +5,9 @@ import * as React from 'react';
 
 import '@/styles/globals.css';
 
-import GoogleAnalytics from '@/components/ga/GoogleAnalytics';
-
 import { siteConfig } from '@/constant/config';
 import { GA_MEASUREMENT_ID, GOOGLE_TAG_MANGER_ID } from '@/constant/env';
+import GoogleAnalytics from '@/features/ga-ab-test/GoogleAnalytics';
 import QueryClientProvider from '@/providers/QueryClientProvider';
 
 export const metadata: Metadata = {
